requirements.txt
dash
chardet
numpy
Pillow
